+++ game_patch/misc/wndproc.cpp
@@ -1,6 +1,7 @@
 #include "wndproc.h"
 #include "../rf/misc.h"
 #include "../rf/network.h"
+#include "../rf/input.h"
 #include <windows.h>
 #include <patch_common/AsmWriter.h>